fix build errors with latest rules_go
Recent rules_go versions do not have a `url` parameter on `go_repository`, so
rules_webtesting, unfortunately, has to be changed.

There is also a new requirement that if a `urls` parameter is specified, than a
`tag` or `commit` parameter may not be. So, we fix up the gorilla/mux and
selenium deps.

This patch also upgrades the rules_go external in this WORKSPACE to its current
HEAD.

This does not address the noci tag problem discussed in #147.

Updates #147