Round up the rect's bounds when converting. (#360)
Rather than purely rounding, which can lead to a rectangle that does
not encompass all of the area of the original rectangle, truncate the x1
and y1 values and get the ceiling of the x2 and y2 values. This
effectively "rounds up" the area of the rectangle, ensuring the original
area is always encompassed.