custom_browser: put custom metadata in first place
PR #394 allowed the removal of --flags when merging metadata args, but
only when the removing args are in the later (higher-precedence)
position. `metadata.merge_files` expects higher precedence files
earlier (and reverses them before calling the actual merger program).

This change prepends the user-supplied metadata json (if supplied) to
the list of files to merge.