feat: add support for darwin m1 browsers
`rules_webtesting` currently extracts browser archives as part
of an build action. Additionally, the platform archive is determined
at repository creation using the `repository_ctx.os` property. This
breaks remote build execution, and also makes it impossible to support
darwin ARM64 because the os name is equal regardless of CPU.

This is a good opportunity to move the extraction from a build action
to the repository fetching (improving caching and being more Bazel-idiomatic).
Additionally, with this approach we will be able to select the browser
binaries at configuration time using `select`, allowing for darwin arm64,
and proper remote build execution.

More details can be found in the description of the `platform_archive`
rule.