build: improve release script to create tarball directly (#429)
Improves the release script to create the tarball directly. This
makes it easier to cut new releases as a few manual steps are
now automated. This change also includes some fallback logic
for finding Copybara for Googlers (like it was done previously),
and no longer relies on a release directory to be specified as `$1`
argument to the Bash script.